moved penguin memes to the extensions section; set up new module for nlp functions
